requirements.txt
aiogram>=3.4.0
python-dotenv>=1.0.0
psutil>=5.9.0
